gpu: nvgpu: pmu: fix MISRA 10.3 issues in pmu_ipc
MISRA Rule 10.3 prohibits implicit assignment of objects of narrower
size or essential type. This fixes MISRA 10.3 violations in pmu_ipc.c

JIRA NVGPU-2841

Reviewed-on: https://git-master.nvidia.com/r/2027768
GVS: Gerrit_Virtual_Submit
Tested-by: mobile promotions <svcmobile_promotions@nvidia.com>

+++ drivers/gpu/nvgpu/include/nvgpu/pmu.h
@@ -314,7 +314,7 @@ struct pmu_sequence {
 	 * has been used so far, as the outbound frame is assembled
 	 * (first FB Queue hdr, then CMD, then payloads).
 	 */
-	 u32 buffer_size_used;
+	u16 buffer_size_used;
 	/* offset to out data in the queue element */
 	u16 fbq_out_offset_in_queue_element;
 };
